OS-2790 【MALATA_F4盒子：文件管理器】 外接设备-点击任意目录-返回-看到的现象是从第一个目录再移动到刚开始所点的目录（必现）

+++ app/src/main/java/com/cantv/media/center/ui/directory/PicView.java
@@ -102,7 +102,7 @@ public class PicView extends ImageView {
 
         @Override
         public int hashCode() {
-            return mBitmapTag.hashCode() + (int) (mDstWidth * mDstHeight) << 5;
+            return mBitmapTag.hashCode() + (mDstWidth * mDstHeight) << 5;
         }
 
         @Override
@@ -125,7 +125,6 @@ public class PicView extends ImageView {
     }
 
     private class PicLoadTask {
-
         private String mTaskPicUri = null;
         private PicViewDecoder mTaskPicDecoder = null;
         private File mTaskSavePicAs = null;
@@ -193,7 +192,6 @@ public class PicView extends ImageView {
                             if (picBitmap == null && cacheSlot != null) {
                                 picBitmap = cacheSlot.getValue();
                             }
-//                            if (cacheSlot != null) mPicCache.releaseSlot(cacheSlot);
                             saveBitmap(picBitmap);
                             picBitmap = tuneBitmap(picBitmap);
                             final Bitmap resultBitmap = picBitmap;
